server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
SQLAlchemy==2.0.34
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.6.1
Pillow==10.4.0
orjson==3.10.7
croniter==3.0.3
pytest==8.3.3
pytest-asyncio==0.23.8
freezegun==1.5.1
greenlet==3.0.3
